Fix packaging and failing js tests
