+++ tests/test_generator.py
@@ -38,83 +38,7 @@ class TestGenerator(TestCase):
             },
         }],
         'want_files': {
-            'pinger-v42.js': '''
-/**
-  Juju Pinger API facade version 42.
-  This file has been generated by the generate command in js-libjuju.
-*/
-
-'use strict';
-
-
-class PingerV42 {
-
-  constructor(transport, info) {
-    this._transport = transport;
-    this._info = info;
-    this.version = 42;
-  }
-
-  /**
-    @param {Function} callback Called when the response from Juju is available,
-      the callback receives an error or null if the operation succeeded.
-  */
-  ping(callback) {
-    const params = {};
-    // Prepare the request to the Juju API.
-    const req = {
-      type: 'Pinger',
-      request: 'Ping',
-      version: 42,
-      params: params
-    };
-    // Send the request to the server.
-    this._transport.write(req, (err, resp) => {
-      if (!callback) {
-        return;
-      }
-      if (err) {
-        callback(err, {});
-        return;
-      }
-      callback(null, {});
-    });
-  }
-
-  /**
-    @param {Function} callback Called when the response from Juju is available,
-      the callback receives an error or null if the operation succeeded.
-  */
-  stop(callback) {
-    const params = {};
-    // Prepare the request to the Juju API.
-    const req = {
-      type: 'Pinger',
-      request: 'Stop',
-      version: 42,
-      params: params
-    };
-    // Send the request to the server.
-    this._transport.write(req, (err, resp) => {
-      if (!callback) {
-        return;
-      }
-      if (err) {
-        callback(err, {});
-        return;
-      }
-      callback(null, {});
-    });
-  }
-}
-
-
-const wrappers = require('../wrappers.js');
-if (wrappers.wrapPinger) {
-  PingerV42 = wrappers.wrapPinger(PingerV42);
-}
-
-module.exports = PingerV42;'''[1:]
+            'pinger-v42.js': helpers.read_data('pinger.js'),
         },
     }]
 
@@ -151,4 +75,4 @@ module.exports = PingerV42;'''[1:]
         for filename, want_code in want_files.items():
             with open(os.path.join(out, filename)) as f:
                 got_code = f.read()
-            self.assertEqual(got_code, want_code, got_code)
+            self.assertIn(want_code, got_code)
