+++ tox.ini
@@ -1,10 +1,10 @@
 [tox]
-envlist = py{36}
+envlist = py3
 
 [testenv]
 basepython =
     devenv: python3
-    py36: python3.6
+    py3: python3
 deps =
     check-manifest
     flake8
